Write httomo CoR sweep output to DLS filesystem
When use of the CoR parameter sweep workflow template has stabilised
such that users are reliably able to view the sweep output via the
artifact alone, it would be sensible to not write the intermediate data
that httomo produces to the DLS filesystem. For now though, for
debugging purposes, it's useful to save these intermediate results.